\name{RcmdrPlugin.MAc-package}
\alias{RcmdrPlugin.MAc-package}
\docType{package}
\title{Rcmdr Plug-in for Meta-Analysis with Correlations (MAc)
}
\description{ This package provides an Rcmdr plug-in for the \bold{MAc} package, which is useful for conducting meta-analyses with correlational data.  This Rcmdr plug-in package will run all of the meta-analytic functions from the \bold{MAc} package though a familiar Graphical User Interface (GUI) environment (e.g., SPSS). For more details regarding the \bold{MAc} package, please see the \bold{MAc} link on the CRAN website \url{http://CRAN.R-project.org/package=MAc}
\tabular{ll}{
Package: \tab RcmdrPlugin.MAc-package\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2010-02-14\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{AC Del Re & William T. Hoyt

Maintainer: AC Del Re <acdelre@gmail.com>
}
\references{ AC Del Re & William T. Hoyt (2010). \emph{MAc: Meta-Analysis for Correlations}. R package version 1.0.3.
  http://CRAN.R-project.org/package=MAc
 
}
\keyword{ package }
\seealso{
\bold{MAc} package:  \url{http://CRAN.R-project.org/package=MAc}
}